点赞功能完成

+++ src/main/java/com/hmdp/service/impl/VoucherOrderServiceImpl.java
@@ -192,7 +192,7 @@ public class VoucherOrderServiceImpl extends ServiceImpl<VoucherOrderMapper, Vou
     public Result createVoucherOrder(Long voucherId) {
         // 实现一人一单功能： 一人一单逻辑
         Long userId = UserHolder.getUser().getId();
-        int count = query().eq("user_id", userId).eq("voucher_id", voucherId).count();
+        Long count = query().eq("user_id", userId).eq("voucher_id", voucherId).count();
         if (count > 0) {
             return Result.fail("用户已经购买过一次！");
         }
@@ -225,7 +225,7 @@ public class VoucherOrderServiceImpl extends ServiceImpl<VoucherOrderMapper, Vou
     public void createVoucherOrder2(VoucherOrder voucherOrder) {
         Long userId = voucherOrder.getUserId();
         // 5.1.查询订单
-        int count = query().eq("user_id", userId).eq("voucher_id", voucherOrder.getVoucherId()).count();
+        Long count = query().eq("user_id", userId).eq("voucher_id", voucherOrder.getVoucherId()).count();
         // 5.2.判断是否存在
         if (count > 0) {
             // 用户已经购买过了
